Implementação método findById

+++ src/main/java/io/github/opeterfreitas/todo/domain/Todo.java
@@ -16,7 +16,7 @@ public class Todo implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String titulo;
     private String descricao;
     private LocalDateTime dataParaFinalizar;
@@ -25,7 +25,7 @@ public class Todo implements Serializable {
     public Todo() {
     }
 
-    public Todo(Long id, String titulo, String descricao, LocalDateTime dataParaFinalizar, boolean finalizado) {
+    public Todo(Integer id, String titulo, String descricao, LocalDateTime dataParaFinalizar, boolean finalizado) {
         this.id = id;
         this.titulo = titulo;
         this.descricao = descricao;
@@ -33,11 +33,11 @@ public class Todo implements Serializable {
         this.finalizado = finalizado;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
